+++ crates/fuel-core/src/service/adapters/executor.rs
@@ -13,7 +13,7 @@ impl fuel_core_executor::ports::TransactionsSource for TransactionsSource {
         &self,
         gas_limit: u64,
         transactions_limit: u16,
-        block_transaction_size_limit: u64,
+        block_transaction_size_limit: u32,
     ) -> Vec<MaybeCheckedTransaction> {
         self.txpool
             .select_transactions(
