+++ crates/services/executor/src/ports.rs
@@ -122,7 +122,7 @@ pub trait TransactionsSource {
         &self,
         gas_limit: u64,
         tx_count_limit: u16,
-        block_transaction_size_limit: u64,
+        block_transaction_size_limit: u32,
     ) -> Vec<MaybeCheckedTransaction>;
 }
